Fix Upload Button Logic
