+++ portal-ui/src/common/SecureComponent/SecureComponent.tsx
@@ -15,95 +15,7 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import React, { cloneElement } from "react";
-import get from "lodash/get";
-import { store } from "../../store";
-import { hasAccessToResource } from "./permissions";
-
-export const hasPermission = (
-  resource: string | string[] | undefined,
-  scopes: string[],
-  matchAll?: boolean,
-  containsResource?: boolean
-) => {
-  if (!resource) {
-    return false;
-  }
-  const state = store.getState();
-  const sessionGrants = state.console.session.permissions || {};
-
-  const globalGrants = sessionGrants["arn:aws:s3:::*"] || [];
-  let resources: string[] = [];
-  let resourceGrants: string[] = [];
-  let containsResourceGrants: string[] = [];
-
-  if (resource) {
-    if (Array.isArray(resource)) {
-      resources = [...resources, ...resource];
-    } else {
-      resources.push(resource);
-    }
-
-    // Filter wildcard items
-    const wildcards = Object.keys(sessionGrants).filter(
-      (item) => item.includes("*") && item !== "arn:aws:s3:::*"
-    );
-
-    const getMatchingWildcards = (path: string) => {
-      const items = wildcards.map((element) => {
-        const wildcardItemSection = element.split(":").slice(-1)[0];
-
-        const replaceWildcard = wildcardItemSection
-            .replace("/", "\\/")
-            .replace("\\/*", "($|(\\/.*?))");
-
-        const inRegExp = new RegExp(`${replaceWildcard}$`, "gm");
-
-        if(inRegExp.exec(path)) {
-          return element;
-        }
-
-        return null;
-      });
-
-      return items.filter(itm => itm !== null);
-    };
-
-    resources.forEach((rsItem) => {
-      // Validation against inner paths & wildcards
-      let wildcardRules =getMatchingWildcards(rsItem);
-
-      let wildcardGrants: string[] = [];
-
-      wildcardRules.forEach((rule) => {
-        if(rule) {
-          const wcResources = get(sessionGrants, rule, []);
-          wildcardGrants = [...wildcardGrants, ...wcResources];
-        }
-      });
-
-      const simpleResources = get(sessionGrants, rsItem, []);
-      const s3Resources = get(sessionGrants, `arn:aws:s3:::${rsItem}/*`, []);
-
-      resourceGrants = [...simpleResources, ...s3Resources, ...wildcardGrants];
-
-      if (containsResource) {
-        const matchResource = `arn:aws:s3:::${rsItem}`;
-
-        Object.entries(sessionGrants).forEach(([key, value]) => {
-          if (key.includes(matchResource)) {
-            containsResourceGrants = [...containsResourceGrants, ...value];
-          }
-        });
-      }
-    });
-  }
-
-  return hasAccessToResource(
-    [...resourceGrants, ...globalGrants, ...containsResourceGrants],
-    scopes,
-    matchAll
-  );
-};
+import hasPermission from "./accessControl";
 
 interface ISecureComponentProps {
   errorProps?: any;
