+++ portal-ui/src/screens/Console/Policies/PolicyDetails.tsx
@@ -51,9 +51,10 @@ import {
   IAM_PAGES,
   IAM_SCOPES,
 } from "../../../common/SecureComponent/permissions";
-import SecureComponent, {
+import {
+  SecureComponent,
   hasPermission,
-} from "../../../common/SecureComponent/SecureComponent";
+} from "../../../common/SecureComponent";
 
 import withSuspense from "../Common/Components/withSuspense";
 import { AppState } from "../../../store";
