+++ portal-ui/src/screens/Console/Users/ListUsers.tsx
@@ -47,10 +47,12 @@ import {
   IAM_SCOPES,
   S3_ALL_RESOURCES,
 } from "../../../common/SecureComponent/permissions";
-import SecureComponent, {
-  hasPermission,
-} from "../../../common/SecureComponent/SecureComponent";
+
 import RBIconButton from "../Buckets/BucketDetails/SummaryItems/RBIconButton";
+import {
+  hasPermission,
+  SecureComponent,
+} from "../../../common/SecureComponent";
 
 const AddUser = withSuspense(React.lazy(() => import("./AddUser")));
 const SetPolicy = withSuspense(
